requirements.txt
qrcode[pil]~=8.0
python-dotenv~=1.1.0
flask
pyotp~=2.9.0
pyzbar~=0.1.9
opencv-python~=4.11.0.86